requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-multipart==0.0.12
jinja2==3.1.4
pydantic==2.9.0
PyYAML==6.0.2
sentence-transformers==3.0.1
torch>=2.2.0
openai==1.44.0
requests==2.32.3
openai-whisper>=20231117
numba==0.60.0
